Reduce amount of generic type (#17)
This PR is trying to reach a middle step between current status and the status of https://github.com/appannie/react-i18n-jed/pull/16

By removing the second generic type, we actually add extra coverage around the locations where the context cause flow to silently set the type to `any`. (this catched 15 errors on our main repo)

And it'll allow us to easily fill in the generic type manually in order to be 100% sure the types aren't `any`. This will make it easily usable with both flow 0.84 and and 0.89.

```js
export default translate<typeof Comp>(Comp);
```